将userid，topicid，postid数据类型改为long

+++ src/main/java/com/eventiming/form2/Controller/AdminController.java
@@ -7,8 +7,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.math.BigInteger;
-
 @RestController
 public class AdminController {
 
@@ -17,22 +15,22 @@ public class AdminController {
 
 // TODO 更新管理员功能，不同权限的用户操作仍需设计
     @PostMapping("/blockuser")
-    public int blockUser(BigInteger userid){
+    public int blockUser(long userid){
         return adminService.blockUser(userid);
     }
     @PostMapping("/deleteuser")
-    public int deleteUser(BigInteger userid){
+    public int deleteUser(long userid){
         return adminService.deleteUser(userid);
     }
 
     @PostMapping("/deleteTopic")
-    public int deleteTopic(BigInteger topicid){
+    public int deleteTopic(long topicid){
         return adminService.deleteTopic(topicid);
     }
 
 
     @PostMapping("/deletePost")
-    public int deletePost(BigInteger postid){
+    public int deletePost(long postid){
         return adminService.deletePost(postid);
     }
 }
