+++ src/main/java/com/eventiming/form2/DAO/postDao.java
@@ -4,21 +4,19 @@ import com.eventiming.form2.pojo.post;
 import jakarta.annotation.Resource;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
-
-import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.util.List;
 
 @Mapper
 @Repository
 public interface postDao {
-    int insertPost(BigInteger topicid,
-                   BigInteger userid,
+    int insertPost(long topicid,
+                   long userid,
                    String postcontext,
                    Timestamp timestamp);
-    int deletePostById(BigInteger postid);
-    int updatePost(BigInteger postid , String postcontext);
-    post selectPostById(BigInteger postid);
-    List<post> selectPostsByTopicId(BigInteger topicid);
-    List<post> selectPostsByUserId(BigInteger userid);
+    int deletePostById(long postid);
+    int updatePost(long postid , String postcontext);
+    post selectPostById(long postid);
+    List<post> selectPostsByTopicId(long topicid);
+    List<post> selectPostsByUserId(long userid);
 }
