+++ src/main/java/com/eventiming/form2/DAO/topiccontextDao.java
@@ -3,13 +3,13 @@ package com.eventiming.form2.DAO;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
+
 @Mapper
 @Repository
 public interface topiccontextDao {
-    int insertContext(BigInteger topicid, String text);
-    int deleteContext(BigInteger topicid);
-    int updateContext(BigInteger topicid, String context);
-    String selectContext(BigInteger topicid);
+    int insertContext(long topicid, String text);
+    int deleteContext(long topicid);
+    int updateContext(long topicid, String context);
+    String selectContext(long topicid);
 
 }
