+++ src/main/java/com/eventiming/form2/DAO/userstatusDao.java
@@ -4,7 +4,7 @@ import com.eventiming.form2.pojo.userstatus;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
+
 
 import java.sql.Time;
 import java.sql.Timestamp;
@@ -13,17 +13,17 @@ import java.util.List;
 @Repository
 public interface userstatusDao {
 
-    int insertUserStatus(BigInteger userid, int usergroup, Timestamp lastactivetime ,int topicnum, int threadnum);
+    int insertUserStatus(long userid, int usergroup, Timestamp lastactivetime ,int topicnum, int threadnum);
 //   1. 根据用户id返回基本信息
-    userstatus selectUserStatusById(BigInteger userid);
+    userstatus selectUserStatusById(long userid);
 //	2. 根据用户id返回用户组
-    int selectUserGroupById(BigInteger userid);
-    int updateUserGroupById(BigInteger userid, int usergroup);
-    int updateUserTopicNum(BigInteger userid, Timestamp timestamp);
-    int updateUserThreadNum(BigInteger userid, Timestamp timestamp);
-    List<BigInteger> selectUserIdByGroup(int group);
-    List<BigInteger> selectUserIdByTopicNum(int min, int max);
-    List<BigInteger> selectUserIdByThreadNum(int min, int max);
-    List<BigInteger> selectUserIdBytopTopicNum(int num, int offset);
-    List<BigInteger> selectUserIdBytopThreadNum(int num, int offset);
+    int selectUserGroupById(long userid);
+    int updateUserGroupById(long userid, int usergroup);
+    int updateUserTopicNum(long userid, Timestamp timestamp);
+    int updateUserThreadNum(long userid, Timestamp timestamp);
+    List<Long> selectUserIdByGroup(int group);
+    List<Long> selectUserIdByTopicNum(int min, int max);
+    List<Long> selectUserIdByThreadNum(int min, int max);
+    List<Long> selectUserIdBytopTopicNum(int num, int offset);
+    List<Long> selectUserIdBytopThreadNum(int num, int offset);
 }
