+++ src/main/java/com/eventiming/form2/Service/PostServiceImpl.java
@@ -4,8 +4,6 @@ import com.eventiming.form2.DAO.postDao;
 import com.eventiming.form2.pojo.post;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -13,24 +11,24 @@ import java.util.List;
 public class PostServiceImpl implements PostService{
     @Autowired
     private postDao postdao;
-    public int addpost(BigInteger userid, BigInteger topicid, String postcontext){
+    public int addpost(long userid, long topicid, String postcontext){
         Timestamp timestamp = new Timestamp(System.currentTimeMillis());
         postdao.insertPost(topicid,userid,postcontext,timestamp);
         return 1;
     }
 
-    public int deletepost(BigInteger postid){
+    public int deletepost(long postid){
         postdao.deletePostById(postid);
         return 0;
     }
 
     // TODO 支持更改
 
-    public List<post> selectPostByTopic(BigInteger topicid){
+    public List<post> selectPostByTopic(long topicid){
         return postdao.selectPostsByTopicId(topicid);
     }
 
-    public List<post> selectPostByUser(BigInteger userid){
+    public List<post> selectPostByUser(long userid){
         return postdao.selectPostsByUserId(userid);
     }
 }
