+++ src/main/java/com/eventiming/form2/pojo/post.java
@@ -4,17 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
-import java.sql.Time;
 import java.sql.Timestamp;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class post {
-    BigInteger postid;
-    BigInteger topicid;
-    BigInteger userid;
+    long postid;
+    long topicid;
+    long userid;
     String postcontext;
     Timestamp timestamp;
