+++ src/main/java/com/eventiming/form2/pojo/user.java
@@ -4,14 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
+
 import java.sql.Timestamp;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class user {
-    BigInteger userid;
+    long userid;
     String username;
     String password;
     String email;
